# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Library API"
  description: "API for managing members, books, and loans."
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
...
paths:
...
  /loans:
    get:
      tags:
        - "Loans"
      summary: "Get all loans"
      responses:
        200:
          description: "Loans retrieved"
          schema:
            $ref: "#/definitions/LoansList"

    post:
      tags:
        - "Loans"
      summary: "Create loan"
      description: "Creates a loan for a member and optional book IDs."
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Loan payload"
          schema:
            $ref: "#/definitions/LoanCreatePayload"
      responses:
        201:
          description: "Loan created"
          schema:
            $ref: "#/definitions/Loan"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Member not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /loans/{loan_id}:
    get:
      tags:
        - "Loans"
      summary: "Get loan by ID"
      parameters:
        - in: "path"
          name: "loan_id"
          required: true
          type: "integer"
      responses:
        200:
          description: "Loan retrieved"
          schema:
            $ref: "#/definitions/Loan"
        404:
          description: "Loan not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      tags:
        - "Loans"
      summary: "Edit books in loan"
      description: "Adds and/or removes books from an existing loan."
      parameters:
        - in: "path"
          name: "loan_id"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          required: true
          description: "Book IDs to add/remove"
          schema:
            $ref: "#/definitions/LoanEditPayload"
      responses:
        200:
          description: "Loan updated"
          schema:
            $ref: "#/definitions/Loan"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Loan not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - "Loans"
      summary: "Delete loan by ID"
      parameters:
        - in: "path"
          name: "loan_id"
          required: true
          type: "integer"
      responses:
        200:
          description: "Loan deleted"
          schema:
            $ref: "#/definitions/DeleteMessage"
        404:
          description: "Loan not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
  ErrorResponse:
    type: "object"
    properties:
      error:
        type: "string"
      messages:
        type: "string"
  DeleteMessage:
    type: "object"
    properties:
      message:
        type: "string"
...
  LoanCreatePayload:
    type: "object"
    required:
      - "loan_date"
      - "member_id"
    properties:
      loan_date:
        type: "string"
        format: "date"
      member_id:
        type: "integer"
      book_ids:
        type: "array"
        items:
          type: "integer"
  LoanEditPayload:
    type: "object"
    required:
      - "add_book_ids"
      - "remove_book_ids"
    properties:
      add_book_ids:
        type: "array"
        items:
          type: "integer"
      remove_book_ids:
        type: "array"
        items:
          type: "integer"
  Loan:
    type: "object"
    properties:
      id:
        type: "integer"
      loan_date:
        type: "string"
        format: "date"
      member_id:
        type: "integer"
      books:
        type: "array"
        items:
          type: "integer"
  LoansList:
    type: "array"
    items:
      $ref: "#/definitions/Loan"